backend/requirements.txt
# FLASK BASE MODULES
flask
flask_cors
requests

# DATABASE MODULES
flask-sqlalchemy
psycopg2

# AUTHENTICATION MODULES
PyJWT

# SPECIFIC MODULES
pandas

google-api-python-client
google-auth-httplib2
google-auth-oauthlib